CareerBot/cry_backend/requirements.txt
aiosmtplib==4.0.2
bcrypt==5.0.0
fastapi==0.118.0
Jinja2==3.1.6
langchain==0.3.27
langserve
langchain-core==0.3.76
langchain-community==0.3.9
langchain-text-splitters==0.3.11
motor==3.7.1
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
prometheus-client==0.23.1
pydantic-settings==2.11.0
pydantic==2.11.9
PyJWT==2.10.1
pymongo==4.15.1
pytest-asyncio==1.2.0
pytest==8.4.2
pytest-timeout==2.2.0
python-dotenv==1.1.1
sse-starlette==3.0.2
structlog==25.4.0
tqdm==4.67.1
uvicorn==0.37.0
openai
langchain-openai
httpx
anyio
rich
websockets
